requirements.txt
python-dateutil==2.5.1
pytest==2.8.3
pytest-django==2.9.1
pytest-cov==2.2.0